+++ crates/hydebar-core/src/modules/keyboard_layout.rs
@@ -75,7 +75,7 @@ impl KeyboardLayout {
 
 impl Module for KeyboardLayout {
     type ViewData<'a> = &'a KeyboardLayoutModuleConfig;
-    type SubscriptionData<'a> = ();
+    type RegistrationData<'a> = ();
 
     fn view(
         &self,
@@ -97,7 +97,7 @@ impl Module for KeyboardLayout {
         }
     }
 
-    fn subscription(&self, _: Self::SubscriptionData<'_>) -> Option<Subscription<app::Message>> {
+    fn subscription(&self) -> Option<Subscription<app::Message>> {
         let id = TypeId::of::<Self>();
 
         let hyprland = Arc::clone(&self.hyprland);